Add files via upload
Another LRAD version. This one is for Arduino MKR Zero and it is not triggered by Computer Vision. It will start as soon as the Arduino starts, unless the potentiometer is in low position. 4 wavs must be placed into the microSD card. The first one lrad0.wav Then lradX.wav